Improve organization of file lock tests
