+++ packages/php-wasm/node/src/test/php-file-locking.spec.ts
@@ -1,7 +1,12 @@
-import { existsSync, mkdtempSync, rmSync, writeFileSync } from 'fs';
+import { mkdtempSync, rmSync, writeFileSync } from 'fs';
 import { join } from 'path';
 import { tmpdir } from 'os';
-import { PHP } from '@php-wasm/universal';
+import {
+	getLoadedRuntime,
+	PHP,
+	proxyFileSystem,
+	type SupportedPHPVersion,
+} from '@php-wasm/universal';
 import { SupportedPHPVersions } from '@php-wasm/universal';
 import {
 	createNodeFsMountHandler,
@@ -9,12 +14,19 @@ import {
 	loadNodeRuntime,
 } from '../lib';
 import {
+	joinPaths,
 	wrapSynchronousInterfaceAsPromised,
 	type Promised,
 } from '@php-wasm/util';
 import { jspi } from 'wasm-feature-detect';
+import type { FileLockManager } from '../lib/file-lock-manager';
 
-describe.each(SupportedPHPVersions)('PHP %s: File locking', (phpVersion) => {
+const phpVersionsToTest =
+	'PHP' in process.env
+		? [process.env['PHP']! as SupportedPHPVersion]
+		: SupportedPHPVersions;
+
+describe.each(phpVersionsToTest)('PHP %s: File locking', (phpVersion) => {
 	const vfsMountPoint = '/test';
 
 	let tempDir: string;
@@ -1388,4 +1400,169 @@ error_log = ${errorLogPath}
 			expect(result2Data.attempt_after_exit).toBe(true);
 		});
 	}, 5000);
+
+	describe(`Additional tests with multiple php-wasm instances`, async () => {
+		const mockFnWithResult = (await jspi())
+			? // Use async mocks for JSPI to match the async FileLockManager
+				// used by JSPI PHP builds.
+				(value: any) => vi.fn().mockResolvedValue(value)
+			: (value: any) => vi.fn().mockReturnValue(value);
+
+		function createMockFileLockManager(): FileLockManager {
+			return {
+				lockWholeFile: mockFnWithResult(true),
+				lockFileByteRange: mockFnWithResult(true),
+				findFirstConflictingByteRangeLock: mockFnWithResult(undefined),
+				releaseLocksForProcessFd: mockFnWithResult(undefined),
+				releaseLocksForProcess: mockFnWithResult(undefined),
+			};
+		}
+
+		// TODO: Add tests for fcntl()
+
+		test(`should attempt to lock a NODEFS file and a PROXYFS node that wraps a NODEFS file`, async () => {
+			// NOTE: Normally, we would use a single file lock manager across all runtimes,
+			// but to keep state clearer within this test, we use a separate manager per runtime.
+			const fileLockManagerForRuntime1 = createMockFileLockManager();
+			const ENV = { DOCROOT: '/wordpress' };
+			const php1 = new PHP(
+				await loadNodeRuntime(phpVersion, {
+					emscriptenOptions: {
+						ENV,
+						fileLockManager: fileLockManagerForRuntime1,
+					},
+				})
+			);
+			const realPathToMount = joinPaths(
+				import.meta.dirname,
+				'test-data',
+				'file-lock-test'
+			);
+			php1.mount('/wordpress', createNodeFsMountHandler(realPathToMount));
+			const realPathToLock = joinPaths(
+				realPathToMount,
+				'wp-content',
+				'lock-this.txt'
+			);
+			const vfsPathToLock = '/wordpress/wp-content/lock-this.txt';
+			const phpThatAttemptsToLock = `<?php
+			$f = fopen('${vfsPathToLock}', 'w');
+			flock($f, LOCK_EX);
+			`;
+			const result1 = await php1.runStream({
+				code: phpThatAttemptsToLock,
+			});
+			expect(await result1.exitCode).toBe(0);
+			expect(
+				fileLockManagerForRuntime1.lockWholeFile,
+				'locking NODEFS file'
+			).toHaveBeenCalledWith(
+				realPathToLock,
+				expect.objectContaining({ type: 'exclusive' })
+			);
+
+			const fileLockManagerForRuntime2 = createMockFileLockManager();
+			const php2 = new PHP(
+				await loadNodeRuntime(phpVersion, {
+					emscriptenOptions: {
+						ENV,
+						fileLockManager: fileLockManagerForRuntime2,
+					},
+				})
+			);
+			proxyFileSystem(php1, php2, ['/wordpress']);
+			const result2 = await php2.runStream({
+				code: phpThatAttemptsToLock,
+			});
+			expect(await result2.exitCode).toBe(0);
+			expect(
+				fileLockManagerForRuntime2.lockWholeFile,
+				'locking NODEFS file via PROXYFS'
+			).toHaveBeenCalledWith(
+				realPathToLock,
+				expect.objectContaining({ type: 'exclusive' })
+			);
+		});
+
+		test(`should not attempt to lock a MEMFS file or a PROXYFS node that wraps a MEMFS file`, async () => {
+			// NOTE: Normally, we would use a single file lock manager across all runtimes,
+			// but to keep state clearer within this test, we use a separate manager per runtime.
+			const fileLockManagerForRuntime1 = createMockFileLockManager();
+			const ENV = { DOCROOT: '/wordpress' };
+			const php1 = new PHP(
+				await loadNodeRuntime(phpVersion, {
+					emscriptenOptions: {
+						ENV,
+						fileLockManager: fileLockManagerForRuntime1,
+					},
+				})
+			);
+			php1.mkdir('/wordpress/wp-content');
+			const pathNotToLock = '/wordpress/wp-content/do-not-lock-this.txt';
+			php1.writeFile(pathNotToLock, new Uint8Array(0));
+			const phpThatAttemptsToLock = `<?php
+				$f = fopen('${pathNotToLock}', 'w');
+				// Explicitly fail so this test does not pass by accident
+				// if the PHP fails to open the file and tolerates the error.
+				if ($f === false) {
+					throw new Error('Failed to open file');
+				}
+				flock($f, LOCK_EX);
+				`;
+			const result1 = await php1.runStream({
+				code: phpThatAttemptsToLock,
+			});
+			expect(await result1.exitCode).toBe(0);
+			expect(
+				fileLockManagerForRuntime1.lockWholeFile
+			).not.toHaveBeenCalled();
+
+			const fileLockManagerForRuntime2 = createMockFileLockManager();
+			const php2 = new PHP(
+				await loadNodeRuntime(phpVersion, {
+					emscriptenOptions: {
+						ENV,
+						fileLockManager: fileLockManagerForRuntime2,
+					},
+				})
+			);
+			proxyFileSystem(php1, php2, ['/wordpress']);
+			const result2 = await php2.runStream({
+				code: phpThatAttemptsToLock,
+			});
+			expect(await result2.exitCode).toBe(0);
+			expect(
+				fileLockManagerForRuntime2.lockWholeFile
+			).not.toHaveBeenCalled();
+		});
+
+		test(`regression test for https://github.com/WordPress/wordpress-playground/pull/2300`, async () => {
+			const opts = {
+				emscriptenOptions: { ENV: { DOCROOT: '/wordpress' } },
+			};
+			const runtime1 = getLoadedRuntime(
+				await loadNodeRuntime('8.3', opts)
+			);
+			runtime1.FS.mkdir('/wordpress');
+
+			const runtime2 = getLoadedRuntime(
+				await loadNodeRuntime('8.3', opts)
+			);
+			runtime2.FS.mkdir('/wordpress');
+
+			runtime2.FS.mount(
+				runtime2.PROXYFS,
+				{ root: '/wordpress', fs: runtime1.FS },
+				'/wordpress'
+			);
+
+			// This worked:
+			// runtime1.FS.mkdir('/wordpress/wp-content');
+
+			// Prior to a fix, this did not:
+			expect(() =>
+				runtime2.FS.mkdir('/wordpress/wp-content')
+			).not.toThrow();
+		});
+	});
 });
